requirements.txt
sqlalchemy
asyncpg
pydantic-settings
fastapi
alembic
bcrypt
python-jose
python-multipart
pydantic[email]
uvicorn
jinja2
httpx